% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_matrix.R
\name{create_interaction_matrix}
\alias{create_interaction_matrix}
\title{wrapper function to use RcppArmadillo to create interaction matrix}
\usage{
create_interaction_matrix(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
interaction matrix
}
\description{
wrapper function to use RcppArmadillo to create interaction matrix
}
