% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{function to convert read data into per second score}
\usage{
extract_data(dataset, location_map, review_data = FALSE)
}
\arguments{
\item{dataset}{data set}

\item{location_map}{map to transcode transponders to ponds}

\item{review_data}{flag to indicate what experiment we are transcoding}
}
\description{
function to convert read data into per second score
}
