% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_to_seconds.R
\name{extract_location_at_all_seconds}
\alias{extract_location_at_all_seconds}
\title{function to calculate location at all times during a single day}
\usage{
extract_location_at_all_seconds(paired_reads)
}
\arguments{
\item{paired_reads}{paired reads output}
}
\value{
tibble with two columns: time and location
}
\description{
function to calculate location at all times during a single day
}
