% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_breeding_spot.R
\name{filter_breeding_spot}
\alias{filter_breeding_spot}
\title{function to filter out breeding spot antennae}
\usage{
filter_breeding_spot(dataset)
}
\arguments{
\item{dataset}{input dataset}

\item{return}{filtered dataset}
}
\description{
This function filters out recordings made at the breeding spot antennae,
being: 30, 40, 50, 60, 70, 80, 90, 100, but also:
10, 15, 25, 45.
Hence, this filters out all reads done by antennae that are not placed at
corridors.
}
