% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fuzzy_match.R
\name{find_fuzzy_match}
\alias{find_fuzzy_match}
\title{function to find an approximate match to a fish name}
\usage{
find_fuzzy_match(focal_fish, fish_names)
}
\arguments{
\item{focal_fish}{focal fish tag}

\item{fish_names}{list with 'true' fish tags}
}
\value{
either the approximate match, or it's own name in case no match is
found
}
\description{
function to find an approximate match to a fish name
}
