% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_paired_Reads.R
\name{find_paired_reads}
\alias{find_paired_reads}
\title{function to extract paired reads that signify a change in container}
\usage{
find_paired_reads(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
tibble with extracted information
}
\description{
function to extract paired reads that signify a change in container
}
