% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_location.R
\name{infer_location}
\alias{infer_location}
\title{function to infer the movement based on antennae reads}
\usage{
infer_location(first_read, second_read)
}
\arguments{
\item{first_read}{first antennae recording}

\item{second_read}{matching second antennae recording}
}
\value{
list with 'old_location' and 'new_location', indicating where the
fish came from and where it moved to
}
\description{
function to infer the movement based on antennae reads
}
