% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_closest.R
\name{pick_closest}
\alias{pick_closest}
\title{function to infer per second the nearest location}
\usage{
pick_closest(
  dataset,
  prev_day_location = NA,
  used_site_map,
  start_time,
  end_time
)
}
\arguments{
\item{dataset}{dataset having at least columns 'Date', 'Time', 'Unit number',
and 'Transponder code'}

\item{prev_day_location}{location of fish at previous day (if known)}

\item{used_site_map}{site map, result of load(site_map) or load(review_map)}

\item{start_time}{starting time of recording}

\item{end_time}{ending time of recording}
}
\value{
tibble
}
\description{
function to infer per second the nearest location
}
