% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_to_seconds.R
\name{reconstruct_location}
\alias{reconstruct_location}
\title{sometimes, fish don't move, so we have to infer where they are from
edge reads}
\usage{
reconstruct_location(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
tibble with two columns: time and location
}
\description{
sometimes, fish don't move, so we have to infer where they are from
edge reads
}
