% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_reads.R
\name{remove_duplicate_reads}
\alias{remove_duplicate_reads}
\title{function remove reads of a fish moving in front of an entrance antenna}
\usage{
remove_duplicate_reads(dataset)
}
\arguments{
\item{dataset}{dataset tibble}
}
\value{
filtered dataset
}
\description{
function to remove duplicate entries
}
