% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_map_data.R
\docType{data}
\name{site_map}
\alias{site_map}
\title{map identifying all antennae per location}
\format{
a tibble
}
\usage{
site_map
}
\description{
a table that can be used to translate antennae reads into
locations
}
\keyword{datasets}
