% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_all_locations_cpp}
\alias{get_all_locations_cpp}
\title{function to get all locations}
\usage{
get_all_locations_cpp(max_sec, times, antennae, site_map, prev_day_location)
}
\arguments{
\item{max_sec}{maximum number of seconds}

\item{times}{time points}

\item{antennae}{vector of antennae}

\item{site_map}{site map}

\item{prev_day_location}{last known location at the previous day.}
}
\value{
vector of sites
}
\description{
function to get all locations
}
