% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_all_locations_cpp}
\alias{get_all_locations_cpp}
\title{function to get all locations}
\usage{
get_all_locations_cpp(max_sec, times, antennae, site_map, prev_day_location)
}
\arguments{
\item{max_sec}{maxsec}

\item{times}{tt}

\item{antennae}{aa}

\item{site_map}{sm}
}
\value{
vector of sites
}
\description{
function to get all locations
}
