% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_matrix_class}
\alias{make_matrix_class}
\title{function to get all locations}
\usage{
make_matrix_class(times, location, fish, num_fish)
}
\arguments{
\item{times}{vector of times}

\item{location}{vector of locations}

\item{fish}{vector of fish}

\item{num_fish}{number of fish}
}
\value{
matrix
}
\description{
function to get all locations
}
