% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_closest.R
\name{pick_closest}
\alias{pick_closest}
\title{function to infer per second the nearest location}
\usage{
pick_closest(dataset, prev_day_location)
}
\arguments{
\item{dataset}{dataset having at least columns 'Date', 'Time', 'Unit number',
and 'Transponder code'}
}
\value{
tibble
}
\description{
function to infer per second the nearest location
}
